+++ src/app/e2e-testee/rate/demo.module.ts
@@ -1,8 +1,8 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {RateDemoComponent} from './rate/rate-full-demo.component';
+import {RateDemoComponent} from './rate/app.component';
 import {RouterModule} from "@angular/router";
-import {RateFullModule} from "../../live-demo/rate/rate/rate-full.module";
+import {RateFullModule} from "../../live-demo/rate/rate/app.module";
 
 const rateDemoRoutes = [
     {path: 'rate-full', component: RateDemoComponent}
